Reorganize code
Game logic moved to Core namespace to separate it from the CLI logic. Some kind of n-layered architecture
