+++ Assignments/Assignment1/Core/GameProcessor.cs
@@ -1,4 +1,4 @@
-﻿namespace Assignment1;
+﻿namespace Assignment1.Core;
 
 internal class GameProcessor(
     UserInputProvider inputProvider,
