+++ Assignments/Assignment1/Core/GameResult.cs
@@ -1,4 +1,4 @@
-﻿namespace Assignment1;
+﻿namespace Assignment1.Core;
 
 internal record struct GameResult(
     PlayerDescriptor? Winner);
