+++ Assignments/Assignment1/Core/LeaderboardRecord.cs
@@ -1,4 +1,4 @@
-﻿namespace Assignment1;
+﻿namespace Assignment1.Core;
 
 internal record LeaderboardRecord(
     string Player,
