+++ Assignments/Assignment1/Core/MoveDescriptor.cs
@@ -1,4 +1,4 @@
-﻿namespace Assignment1;
+﻿namespace Assignment1.Core;
 
 internal record MoveDescriptor(
     int Row, int Col);
