+++ Assignments/Assignment1/Core/PlayersProvider.cs
@@ -1,4 +1,4 @@
-﻿namespace Assignment1;
+﻿namespace Assignment1.Core;
 
 internal class PlayersProvider
 {
